requirements.txt
APScheduler==3.2.0
beautifulsoup4==4.4.1
bs4==0.0.1
click==6.6
cssselect==0.9.2
feedfinder2==0.0.4
feedparser==5.2.1
Flask==0.11.1
Flask-SQLAlchemy==2.1
idna==2.1
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.8
lxml==3.6.0
MarkupSafe==0.23
newspaper3k==0.1.7
nltk==3.2.1
Pillow==3.3.0
python-dateutil==2.5.3
python-slugify==1.2.0
pytz==2016.6.1
PyYAML==3.11
requests==2.10.0
requests-file==1.4
six==1.10.0
SQLAlchemy==1.0.14
tldextract==2.0.1
tzlocal==1.2.2
Unidecode==0.4.19
Werkzeug==0.11.10
